# Reduce allocations by pooling buffered JSON objects